+++ projects/ks89/angular-modal-gallery/src/lib/model/preview-config.interface.ts
@@ -29,6 +29,7 @@ import { Size } from './size.interface';
  */
 export interface PreviewConfig {
   visible: boolean;
+  // by default previews are hidden on mobile
   mobileVisible?: boolean;
   number?: number;
   arrows?: boolean;